refactor: move PlaceOrderButton to PlaceOrderButtonPaymentSource
Pure rename ahead of the Payments Model split. The component's logic is
untouched so the history stays blame-able: its enablement machine is built
on payment_method / payment_source / payment_response / onsubmit refs, none
of which exist on the payment_sessions model.

PlaceOrderButton is a temporary re-export alias, replaced by the router in
the next commit. The existing spec now targets the branch directly, since
its fixtures carry no available_payment_methods and would otherwise resolve
to the undetermined state.

See docs/adr/2026-08-18-place-order-split-by-payments-model.md
