+++ packages/react-components/src/index.ts
@@ -66,6 +66,7 @@ export * from "#components/orders/OrderNumber"
 export * from "#components/orders/OrderStorage"
 export * from "#components/orders/PaymentMethodAmount"
 export * from "#components/orders/PlaceOrderButton"
+export * from "#components/orders/PlaceOrderButtonPaymentSource"
 export * from "#components/orders/PlaceOrderContainer"
 export * from "#components/orders/PrivacyAndTermsCheckbox"
 export * from "#components/orders/ShippingAmount"
